Add NextJS TopLoader and refactor component structure
