+++ package.json
@@ -20,6 +20,7 @@
     "lucide-react": "^0.477.0",
     "negotiator": "^1.0.0",
     "next": "15.2.1",
+    "nextjs-toploader": "^3.7.15",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hot-toast": "^2.5.2",
