--- src/app/components/common/welcomecard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './WelcomeCard';
