+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import "./globals.css";
 import { defaultLocale } from "@/i18n/config";
 import '@/styles/main.scss';
-
+import NextTopLoader from 'nextjs-toploader';
 export const metadata: Metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
@@ -20,8 +20,9 @@ export default async function RootLayout({
   const lang = resolvedParams.lang || defaultLocale;
   
   return (
-    <html lang={lang}>
+    <html lang={lang} suppressHydrationWarning>
       <body>
+        <NextTopLoader />
         {children}
       </body>
     </html>
