+++ src/components/common/welcomecard/WelcomeCard.scss
